+++ src/autoskillit/recipe/rules/rules_ci.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 import regex as re
 
-from autoskillit.core import PRState, Severity
-from autoskillit.recipe._analysis import ValidationContext, _extract_routing_edges
+from autoskillit.core import Severity
+from autoskillit.recipe._analysis import ValidationContext
 from autoskillit.recipe.registry import RuleFinding, semantic_rule
 
 _NO_RUNS_RE = re.compile(r"""==\s*['"]?no_runs['"]?""")
@@ -66,158 +66,6 @@ def _check_ci_polling_inline_shell(ctx: ValidationContext) -> list[RuleFinding]:
                     ),
                 )
             )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# ci-failure-missing-conflict-gate helpers
-# ---------------------------------------------------------------------------
-
-_CONFLICT_GATE_KEYWORDS: frozenset[str] = frozenset({"merge-base", "is-ancestor"})
-_CONFLICT_RESOLUTION_SKILLS: frozenset[str] = frozenset({"resolve-merge-conflicts"})
-_CODE_RESOLUTION_SKILLS: frozenset[str] = frozenset({"resolve-failures"})
-
-
-def _is_exit_code_conflict_gate(step: object) -> bool:
-    """Return True if step is a run_cmd stale-base exit-code gate (merge-base/is-ancestor)."""
-    if getattr(step, "tool", None) != "run_cmd":
-        return False
-    cmd = (getattr(step, "with_args", {}) or {}).get("cmd", "")
-    return isinstance(cmd, str) and any(kw in cmd for kw in _CONFLICT_GATE_KEYWORDS)
-
-
-def _is_conflict_resolution_step(step: object) -> bool:
-    """Return True if step invokes resolve-merge-conflicts via run_skill."""
-    if getattr(step, "tool", None) != "run_skill":
-        return False
-    skill_cmd = (getattr(step, "with_args", {}) or {}).get("skill_command", "")
-    return isinstance(skill_cmd, str) and any(s in skill_cmd for s in _CONFLICT_RESOLUTION_SKILLS)
-
-
-def _is_code_resolution_step(step: object) -> bool:
-    """Return True if step invokes code-level CI resolution (resolve-failures)."""
-    if getattr(step, "tool", None) != "run_skill":
-        return False
-    with_args = getattr(step, "with_args", {}) or {}
-    skill_cmd = with_args.get("skill_command", "")
-    return isinstance(skill_cmd, str) and any(s in skill_cmd for s in _CODE_RESOLUTION_SKILLS)
-
-
-def _bfs_without_barrier(graph: dict[str, set[str]], start: str, barriers: set[str]) -> set[str]:
-    """BFS from start; barrier nodes are visited but not expanded."""
-    reachable: set[str] = set()
-    queue = [start]
-    while queue:
-        node = queue.pop()
-        if node in reachable:
-            continue
-        reachable.add(node)
-        if node in barriers:
-            continue
-        for successor in graph.get(node, set()):
-            if successor not in reachable:
-                queue.append(successor)
-    return reachable
-
-
-@semantic_rule(
-    name="ci-failure-missing-conflict-gate",
-    description=(
-        "wait_for_ci failure route reaches resolve-failures without a stale-base "
-        "detection gate (run_cmd merge-base check or resolve-merge-conflicts)"
-    ),
-    severity=Severity.ERROR,
-)
-def _check_ci_failure_conflict_gate(ctx: ValidationContext) -> list[RuleFinding]:
-    # Identify all conflict-gate and code-resolution steps by name
-    conflict_gates: set[str] = {
-        name
-        for name, step in ctx.recipe.steps.items()
-        if _is_exit_code_conflict_gate(step) or _is_conflict_resolution_step(step)
-    }
-    code_resolution_steps: set[str] = {
-        name for name, step in ctx.recipe.steps.items() if _is_code_resolution_step(step)
-    }
-
-    # If no automated code-resolution loop exists, skip (merge-prs.yaml pattern)
-    if not code_resolution_steps:
-        return []
-
-    findings: list[RuleFinding] = []
-    for name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_ci":
-            continue
-        failure_target = step.on_failure
-        if failure_target is None:
-            continue
-
-        # BFS from the failure target; conflict-gate steps are barriers
-        reachable = _bfs_without_barrier(ctx.step_graph, failure_target, conflict_gates)
-
-        # If any code-resolution step is reachable before a conflict gate → violation
-        unguarded = reachable & code_resolution_steps
-        if unguarded:
-            findings.append(
-                RuleFinding(
-                    rule="ci-failure-missing-conflict-gate",
-                    severity=Severity.ERROR,
-                    step_name=name,
-                    message=(
-                        f"Step '{name}' routes CI failures to code-level resolution "
-                        f"({', '.join(sorted(unguarded))}) without a stale-base detection gate. "
-                        "Insert a run_cmd step using 'git merge-base --is-ancestor' (or a "
-                        "resolve-merge-conflicts skill step) before any resolve-failures "
-                        "invocation on the CI failure path."
-                    ),
-                )
-            )
-    return findings
-
-
-@semantic_rule(
-    name="mergeability-conflicting-direct-to-resolution",
-    description=(
-        "check_pr_mergeable CONFLICTING arm reaches a stale-base exit-code gate "
-        "(run_cmd with merge-base) before reaching conflict resolution "
-        "(resolve-merge-conflicts). Route CONFLICTING directly to resolution."
-    ),
-    severity=Severity.ERROR,
-)
-def _check_mergeability_conflicting_direct(ctx: ValidationContext) -> list[RuleFinding]:
-    exit_code_gates: set[str] = {
-        name for name, step in ctx.recipe.steps.items() if _is_exit_code_conflict_gate(step)
-    }
-    resolution_steps: set[str] = {
-        name for name, step in ctx.recipe.steps.items() if _is_conflict_resolution_step(step)
-    }
-    if not exit_code_gates:
-        return []
-    findings: list[RuleFinding] = []
-    for name, step in ctx.recipe.steps.items():
-        if step.tool != "check_pr_mergeable":
-            continue
-        if step.on_result is None:
-            continue
-        for cond in step.on_result.conditions:
-            if not cond.when or "CONFLICTING" not in cond.when:
-                continue
-            reachable = _bfs_without_barrier(ctx.step_graph, cond.route, resolution_steps)
-            unguarded_gates = reachable & exit_code_gates
-            if unguarded_gates:
-                findings.append(
-                    RuleFinding(
-                        rule="mergeability-conflicting-direct-to-resolution",
-                        severity=Severity.ERROR,
-                        step_name=name,
-                        message=(
-                            f"Step '{name}' routes CONFLICTING to '{cond.route}' which "
-                            f"reaches exit-code gate(s) "
-                            f"({', '.join(sorted(unguarded_gates))}) "
-                            f"before conflict resolution. GitHub's CONFLICTING status is "
-                            f"authoritative — route directly to resolve-merge-conflicts."
-                        ),
-                    )
-                )
     return findings
 
 
@@ -364,141 +212,6 @@ def _check_ci_hardcoded_workflow(ctx: ValidationContext) -> list[RuleFinding]:
     return findings
 
 
-# ---------------------------------------------------------------------------
-# wait_for_merge_queue routing rules (I7 + I8)
-# ---------------------------------------------------------------------------
-
-_REQUIRED_MQ_PR_STATES: frozenset[str] = frozenset(
-    s.value for s in PRState if s not in {PRState.ERROR, PRState.NOT_ENROLLED}
-)
-_PR_STATE_WHEN_RE = re.compile(r"\$\{\{\s*result\.pr_state\s*\}\}\s*==\s*(\w+)")
-_MQ_EXPECTED_FALLBACK = "register_clone_unconfirmed"
-
-
-def _extract_mq_when_values(on_result: object) -> set[str]:
-    """Return the set of pr_state values covered by explicit when conditions."""
-    values: set[str] = set()
-    for cond in getattr(on_result, "conditions", []):
-        if getattr(cond, "when", None) is None:
-            continue
-        m = _PR_STATE_WHEN_RE.search(cond.when)
-        if m:
-            values.add(m.group(1))
-    return values
-
-
-def _recipe_has_mq_routing_step(ctx: ValidationContext) -> bool:
-    """Return True if any step uses wait_for_merge_queue with predicate on_result routing."""
-    return any(
-        step.tool == "wait_for_merge_queue"
-        and step.on_result is not None
-        and getattr(step.on_result, "conditions", None)
-        for step in ctx.recipe.steps.values()
-    )
-
-
-def _recipe_uses_register_clone_unconfirmed(ctx: ValidationContext) -> bool:
-    """Return True if this recipe family uses register_clone_unconfirmed as timeout escalation.
-
-    Used only by Rule I8 (conformance targets) — implementation/remediation-family recipes
-    that define register_clone_unconfirmed must route fallback and on_failure there.  Other
-    recipe families (e.g. merge-prs.yaml) route queue timeouts/errors differently and are
-    exempt from target-specific conformance checks, but NOT from PRState completeness (I7).
-    """
-    return _MQ_EXPECTED_FALLBACK in ctx.recipe.steps
-
-
-@semantic_rule(
-    name="wait-for-merge-queue-routing-covers-all-pr-states",
-    description=(
-        "Every non-error PRState value must have an explicit when arm in "
-        "wait_for_merge_queue on_result; prevents silent routing of new states to fallback"
-    ),
-    severity=Severity.ERROR,
-)
-def _check_wait_for_merge_queue_routing_covers_all_pr_states(
-    ctx: ValidationContext,
-) -> list[RuleFinding]:
-    if not _recipe_has_mq_routing_step(ctx):
-        return []
-    findings: list[RuleFinding] = []
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_merge_queue":
-            continue
-        if step.on_result is None or not step.on_result.conditions:
-            continue
-        covered = _extract_mq_when_values(step.on_result)
-        missing = _REQUIRED_MQ_PR_STATES - covered
-        if missing:
-            findings.append(
-                RuleFinding(
-                    rule="wait-for-merge-queue-routing-covers-all-pr-states",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step {step_name!r} is missing explicit routing arms for "
-                        f"PRState values: {sorted(missing)}. Every non-error PRState "
-                        f"must have an explicit when condition in on_result."
-                    ),
-                )
-            )
-    return findings
-
-
-@semantic_rule(
-    name="wait-for-merge-queue-routing-conforms-to-expected-targets",
-    description=(
-        "wait_for_merge_queue fallback and on_failure must both target "
-        "register_clone_unconfirmed; prevents silent success routing on timeout/unknown states"
-    ),
-    severity=Severity.ERROR,
-)
-def _check_wait_for_merge_queue_routing_conforms_to_expected_targets(
-    ctx: ValidationContext,
-) -> list[RuleFinding]:
-    if not _recipe_uses_register_clone_unconfirmed(ctx):
-        return []
-    findings: list[RuleFinding] = []
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_merge_queue":
-            continue
-        # Check fallback condition (when=None)
-        if step.on_result is not None and step.on_result.conditions:
-            fallback_routes = [
-                c.route for c in step.on_result.conditions if getattr(c, "when", None) is None
-            ]
-            for route in fallback_routes:
-                if route != _MQ_EXPECTED_FALLBACK:
-                    findings.append(
-                        RuleFinding(
-                            rule="wait-for-merge-queue-routing-conforms-to-expected-targets",
-                            severity=Severity.ERROR,
-                            step_name=step_name,
-                            message=(
-                                f"Step {step_name!r} has fallback route {route!r} but "
-                                f"expected {_MQ_EXPECTED_FALLBACK!r}. The fallback must "
-                                f"route to register_clone_unconfirmed so unrecognised states "
-                                f"are escalated, not silently treated as success."
-                            ),
-                        )
-                    )
-        # Check on_failure
-        if step.on_failure is not None and step.on_failure != _MQ_EXPECTED_FALLBACK:
-            findings.append(
-                RuleFinding(
-                    rule="wait-for-merge-queue-routing-conforms-to-expected-targets",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step {step_name!r} has on_failure={step.on_failure!r} but "
-                        f"expected {_MQ_EXPECTED_FALLBACK!r}. Tool errors must route "
-                        f"to register_clone_unconfirmed, not a success-path step."
-                    ),
-                )
-            )
-    return findings
-
-
 @semantic_rule(
     name="ci-event-literal-merge-group",
     description=(
@@ -523,69 +236,6 @@ def _check_ci_event_literal_merge_group(ctx: ValidationContext) -> list[RuleFind
                         "wait_for_ci must not hardcode event='merge_group'. "
                         "Use context.ci_event (which is 'push' or null) for pre-queue CI, "
                         "or add a lifecycle-aware condition for in-queue CI."
-                    ),
-                )
-            )
-    return findings
-
-
-_CI_APPLICABLE_RE = re.compile(r"ci_applicable")
-_PRIMARY_CI_EVENT_RE = re.compile(r"context\.(conflict_)?ci_event\s*\}\}")
-
-
-@semantic_rule(
-    name="ci-wait-requires-applicability-guard",
-    description=(
-        "A wait_for_ci step whose event comes from check_repo_merge_state must have "
-        "an upstream action:route step that checks ci_applicable to prevent timeout "
-        "waste when no CI workflows apply."
-    ),
-    severity=Severity.ERROR,
-)
-def _check_ci_wait_requires_applicability_guard(ctx: ValidationContext) -> list[RuleFinding]:
-    from collections import deque
-
-    findings: list[RuleFinding] = []
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_ci":
-            continue
-        event_val = (step.with_args or {}).get("event", "")
-        if not isinstance(event_val, str) or not _PRIMARY_CI_EVENT_RE.search(event_val):
-            continue
-        has_guard = False
-        initial_preds = ctx.predecessors.get(step_name, set())
-        visited: set[str] = set(initial_preds)
-        queue: deque[str] = deque(initial_preds)
-        while queue and not has_guard:
-            node = queue.popleft()
-            pred_step = ctx.recipe.steps.get(node)
-            if pred_step is None:
-                continue
-            if pred_step.tool == "wait_for_ci":
-                has_guard = True
-                break
-            if getattr(pred_step, "action", None) == "route":
-                if pred_step.on_result and pred_step.on_result.conditions:
-                    for cond in pred_step.on_result.conditions:
-                        if cond.when and _CI_APPLICABLE_RE.search(cond.when):
-                            has_guard = True
-                            break
-            if not has_guard:
-                new_preds = ctx.predecessors.get(node, set()) - visited
-                visited.update(new_preds)
-                queue.extend(new_preds)
-        if not has_guard:
-            findings.append(
-                RuleFinding(
-                    rule="ci-wait-requires-applicability-guard",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' calls wait_for_ci with a ci_event from "
-                        f"check_repo_merge_state but has no upstream action:route step "
-                        f"that checks ci_applicable. When ci_applicable=false, "
-                        f"wait_for_ci will exhaust its timeout budget polling for CI runs "
-                        f"that will never appear."
                     ),
                 )
             )
@@ -621,270 +271,6 @@ def _check_ci_timeout_minimum(ctx: ValidationContext) -> list[RuleFinding]:
                     message=(
                         f"timeout_seconds={timeout} is below the {_CI_TIMEOUT_MINIMUM}s "
                         f"minimum — average CI duration is ~317s with peaks to 392s."
-                    ),
-                )
-            )
-    return findings
-
-
-@semantic_rule(
-    name="ci-timed-out-self-loop-unguarded",
-    description=(
-        "A wait_for_ci step whose on_result routes timed_out back to itself "
-        "must have a check_loop_iteration guard on that path to prevent unbounded looping. "
-        "timed_out means CI is still running — it should be polled with a bounded iteration "
-        "count, not routed to a bare self-loop with no cap."
-    ),
-    severity=Severity.ERROR,
-)
-def _check_ci_timed_out_self_loop_unguarded(ctx: ValidationContext) -> list[RuleFinding]:
-    findings: list[RuleFinding] = []
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_ci":
-            continue
-        if not step.on_result or not step.on_result.conditions:
-            continue
-        conditions = step.on_result.conditions
-        timed_out_routes: list[str] = []
-        for cond in conditions:
-            if cond.when is not None and "timed_out" in cond.when:
-                timed_out_routes.append(cond.route)
-        for route in timed_out_routes:
-            if route == step_name:
-                findings.append(
-                    RuleFinding(
-                        rule="ci-timed-out-self-loop-unguarded",
-                        severity=Severity.ERROR,
-                        step_name=step_name,
-                        message=(
-                            f"Step '{step_name}' has a timed_out self-loop with no "
-                            f"check_loop_iteration guard on the path. "
-                            f"Unbounded polling can result from this pattern."
-                        ),
-                    )
-                )
-    return findings
-
-
-@semantic_rule(
-    name="ci-conflict-path-missing-auto-trigger",
-    description=(
-        "A wait_for_ci step on a conflict-resolution path must have auto_trigger: true. "
-        "After a force-push of a conflict fix, CI may not have been triggered yet — "
-        "auto_trigger ensures the empty-commit CI trigger fires on the next poll cycle."
-    ),
-    severity=Severity.ERROR,
-)
-def _check_ci_conflict_path_missing_auto_trigger(ctx: ValidationContext) -> list[RuleFinding]:
-    findings: list[RuleFinding] = []
-    conflict_trigger_steps: set[str] = set()
-    for step_name, step in ctx.recipe.steps.items():
-        if not step.on_result or not step.on_result.conditions:
-            continue
-        for cond in step.on_result.conditions:
-            if cond.when is not None and "CONFLICTING" in cond.when:
-                conflict_trigger_steps.add(step_name)
-                break
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_ci":
-            continue
-        is_on_conflict_path = (
-            "conflict" in step_name.lower()
-            or step_name in conflict_trigger_steps
-            or _has_conflict_ancestor(step_name, conflict_trigger_steps, ctx)
-        )
-        if not is_on_conflict_path:
-            continue
-        auto_trigger = (step.with_args or {}).get("auto_trigger", "")
-        if str(auto_trigger).lower() != "true":
-            findings.append(
-                RuleFinding(
-                    rule="ci-conflict-path-missing-auto-trigger",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' is on a conflict-resolution path but lacks "
-                        f"auto_trigger: 'true' in with_args. After a conflict fix push, "
-                        f"CI may not have fired yet — auto_trigger is mandatory here."
-                    ),
-                )
-            )
-    return findings
-
-
-def _has_conflict_ancestor(
-    step_name: str, conflict_steps: set[str], ctx: ValidationContext
-) -> bool:
-    from collections import deque
-
-    visited: set[str] = set()
-    queue = deque(ctx.predecessors.get(step_name, set()))
-    while queue:
-        current = queue.popleft()
-        if current in visited:
-            continue
-        visited.add(current)
-        if current in conflict_steps:
-            return True
-        queue.extend(ctx.predecessors.get(current, set()))
-    return False
-
-
-@semantic_rule(
-    name="enqueue-missing-ci-gate",
-    description=(
-        "Flags enqueue_pr steps reachable from recipe entry without a wait_for_ci ancestor. "
-        "Premature queue enrollment (before CI passes) causes predictable GitHub rejection."
-    ),
-    severity=Severity.ERROR,
-)
-def _check_enqueue_missing_ci_gate(ctx: ValidationContext) -> list[RuleFinding]:
-    findings: list[RuleFinding] = []
-
-    enqueue_steps: set[str] = {
-        name for name, step in ctx.recipe.steps.items() if step.tool == "enqueue_pr"
-    }
-    if not enqueue_steps:
-        return findings
-
-    ci_gate_steps: set[str] = {
-        name for name, step in ctx.recipe.steps.items() if step.tool == "wait_for_ci"
-    }
-    for name, step in ctx.recipe.steps.items():
-        if getattr(step, "action", None) == "route" and step.on_result:
-            if step.on_result.conditions and any(
-                c.when and _CI_APPLICABLE_RE.search(c.when) for c in step.on_result.conditions
-            ):
-                ci_gate_steps.add(name)
-
-    # Build a direct routing graph without skip_when_false bypass edges.
-    # Bypass edges allow CI gates to be sidestepped in the compiled step_graph,
-    # which would produce false positives here (a wait_for_ci bypassed when
-    # open_pr=false is not a real ungated path to enqueue_pr).
-    step_names = set(ctx.recipe.steps)
-    direct_graph: dict[str, set[str]] = {name: set() for name in step_names}
-    for name, step in ctx.recipe.steps.items():
-        for edge in _extract_routing_edges(step):
-            if edge.edge_type == "exhausted" and step.action is not None:
-                continue
-            if edge.target in step_names:
-                direct_graph[name].add(edge.target)
-
-    # Entry steps: those not pointed to by any step in the direct graph.
-    has_predecessor: set[str] = {s for succs in direct_graph.values() for s in succs}
-    entry_steps = step_names - has_predecessor
-    if not entry_steps:
-        entry_steps = step_names
-
-    # Forward BFS from all entry points; CI gates are barriers — visited but
-    # not expanded. This yields every step reachable from the recipe entry
-    # without every path first crossing a wait_for_ci guard.
-    reachable_without_gate: set[str] = set()
-    queue: list[str] = list(entry_steps)
-    while queue:
-        node = queue.pop()
-        if node in reachable_without_gate:
-            continue
-        reachable_without_gate.add(node)
-        if node in ci_gate_steps:
-            continue
-        for successor in direct_graph.get(node, set()):
-            if successor not in reachable_without_gate:
-                queue.append(successor)
-
-    for enqueue_name in enqueue_steps:
-        if enqueue_name in reachable_without_gate:
-            findings.append(
-                RuleFinding(
-                    rule="enqueue-missing-ci-gate",
-                    severity=Severity.ERROR,
-                    step_name=enqueue_name,
-                    message=(
-                        f"Step '{enqueue_name}' calls enqueue_pr but is reachable from "
-                        "recipe entry without passing through a wait_for_ci step. "
-                        "Add a CI gate (wait_for_ci) before enqueue to prevent premature "
-                        "queue enrollment."
-                    ),
-                )
-            )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# ci-cwd-branch-context-mismatch rule
-# ---------------------------------------------------------------------------
-
-_WORKTREE_BRANCH_PATTERNS: frozenset[str] = frozenset(
-    {
-        "worktree_branch_name",
-        "worktree_branch",
-    }
-)
-
-_CLONE_CWD_PATTERNS: frozenset[str] = frozenset(
-    {
-        "work_dir",
-        "clone_path",
-        "clone_dir",
-    }
-)
-
-_CONTEXT_VAR_RE = re.compile(r"\$\{\{\s*context\.(\w+)\s*\}\}")
-
-
-def _extract_context_var(value: str) -> str | None:
-    m = _CONTEXT_VAR_RE.search(value)
-    return m.group(1) if m else None
-
-
-def _is_worktree_branch_ref(var_name: str) -> bool:
-    return any(pat in var_name for pat in _WORKTREE_BRANCH_PATTERNS)
-
-
-def _is_clone_cwd_ref(var_name: str) -> bool:
-    return any(pat in var_name for pat in _CLONE_CWD_PATTERNS)
-
-
-@semantic_rule(
-    name="ci-cwd-branch-context-mismatch",
-    description=(
-        "wait_for_ci step watches a worktree branch but cwd references the clone root. "
-        "git rev-parse HEAD in the clone root returns the wrong SHA, causing "
-        "_validate_run_matches_scope to silently reject valid CI runs."
-    ),
-    severity=Severity.ERROR,
-)
-def _check_ci_cwd_branch_context_mismatch(ctx: ValidationContext) -> list[RuleFinding]:
-    findings: list[RuleFinding] = []
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "wait_for_ci":
-            continue
-        with_args = step.with_args or {}
-
-        if with_args.get("head_sha"):
-            continue
-
-        branch_val = with_args.get("branch", "")
-        cwd_val = with_args.get("cwd", "")
-
-        branch_var = _extract_context_var(str(branch_val))
-        cwd_var = _extract_context_var(str(cwd_val))
-
-        if not branch_var or not cwd_var:
-            continue
-
-        if _is_worktree_branch_ref(branch_var) and _is_clone_cwd_ref(cwd_var):
-            findings.append(
-                RuleFinding(
-                    rule="ci-cwd-branch-context-mismatch",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' watches worktree branch "
-                        f"(context.{branch_var}) but cwd uses clone root "
-                        f"(context.{cwd_var}). git rev-parse HEAD in the clone root "
-                        f"returns the batch/base branch SHA, not the worktree branch SHA. "
-                        f"Use cwd: '${{{{ context.worktree_path }}}}' or pass head_sha explicitly."
                     ),
                 )
             )
