+++ src/autoskillit/recipe/rules/rules_dataflow.py
@@ -2,18 +2,9 @@
 
 from __future__ import annotations
 
-import importlib
-import inspect
-
-from autoskillit.core import (
-    PIPELINE_FORBIDDEN_TOOLS,
-    SKILL_TOOLS,
-    Severity,
-    get_logger,
-)
+from autoskillit.core import PIPELINE_FORBIDDEN_TOOLS, SKILL_TOOLS, Severity
 from autoskillit.recipe._analysis import ValidationContext
 from autoskillit.recipe.contracts import (
-    _CONTEXT_REF_RE,
     RESULT_CAPTURE_RE,
     get_callable_contract,
     get_skill_contract,
@@ -21,8 +12,6 @@ from autoskillit.recipe.contracts import (
     resolve_skill_name,
 )
 from autoskillit.recipe.registry import RuleFinding, semantic_rule
-
-logger = get_logger(__name__)
 
 
 @semantic_rule(
@@ -228,485 +217,4 @@ def _check_dead_output(ctx: ValidationContext) -> list[RuleFinding]:
                 message=w.message,
             )
         )
-    return findings
-
-
-@semantic_rule(
-    name="implicit-handoff",
-    description="Skill with declared outputs missing capture block",
-    severity=Severity.ERROR,
-)
-def _check_implicit_handoff(ctx: ValidationContext) -> list[RuleFinding]:
-    """Error when a skill step has contract outputs but no capture: block."""
-    wf = ctx.recipe
-    try:
-        manifest = load_bundled_manifest()
-    except Exception:
-        logger.warning("implicit-handoff: failed to load skill_contracts.yaml; skipping")
-        return []
-
-    findings: list[RuleFinding] = []
-    for step_name, step in wf.steps.items():
-        if step.tool not in SKILL_TOOLS:
-            continue
-        skill_cmd = step.with_args.get("skill_command", "")
-        if not isinstance(skill_cmd, str):
-            continue
-        skill_name = resolve_skill_name(skill_cmd)
-        if not skill_name:
-            continue
-        contract = manifest.get("skills", {}).get(skill_name)
-        if not contract:
-            continue
-        if not contract.get("outputs"):
-            continue
-        if not step.capture and not step.capture_list:
-            findings.append(
-                RuleFinding(
-                    rule="implicit-handoff",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' calls '/autoskillit:{skill_name}' "
-                        f"which declares {len(contract['outputs'])} output(s) "
-                        f"but has no capture: or capture_list: block. Add a capture: block to "
-                        f"thread the skill's structured output into pipeline context."
-                    ),
-                )
-            )
-    return findings
-
-
-@semantic_rule(
-    name="multipart-iteration-notes",
-    description="Multi-part plan recipes must declare iteration conventions.",
-    severity=Severity.ERROR,
-)
-def _check_multipart_iteration_notes(ctx: ValidationContext) -> list[RuleFinding]:
-    wf = ctx.recipe
-    _MULTIPART_SKILLS = {"/autoskillit:make-plan", "/autoskillit:rectify"}
-    findings: list[RuleFinding] = []
-
-    has_multipart_step = any(
-        step.tool in SKILL_TOOLS
-        and any(s in step.with_args.get("skill_command", "") for s in _MULTIPART_SKILLS)
-        for step in wf.steps.values()
-    )
-    if not has_multipart_step:
-        return []
-
-    plan_step = wf.steps.get("plan")
-    plan_note = (plan_step.note or "") if plan_step is not None else ""
-    # Also accept the glob pattern from the note of whichever step invokes the multipart skill
-    multipart_step_notes = [
-        (step.note or "")
-        for step in wf.steps.values()
-        if step.tool in SKILL_TOOLS
-        and any(s in step.with_args.get("skill_command", "") for s in _MULTIPART_SKILLS)
-    ]
-    if "*_part_*.md" not in plan_note and not any(
-        "*_part_*.md" in note for note in multipart_step_notes
-    ):
-        findings.append(
-            RuleFinding(
-                rule="multipart-glob-note",
-                severity=Severity.ERROR,
-                step_name="plan",
-                message=(
-                    "Recipe uses make-plan or rectify but neither the 'plan' step note nor "
-                    "the planning step's own note contains '*_part_*.md'. Agents will not "
-                    "know to glob for multi-part plan files. Add: "
-                    "'Glob plan_dir for *_part_*.md or single plan file.' to the plan "
-                    "step's note (or to the make-plan/rectify step's note if no separate "
-                    "'plan' step exists)."
-                ),
-            )
-        )
-
-    sequential_keywords = ("SEQUENTIAL EXECUTION", "full cycle", "Never run verify for all parts")
-    rules_text = " ".join(wf.kitchen_rules)
-    if not any(kw in rules_text for kw in sequential_keywords):
-        findings.append(
-            RuleFinding(
-                rule="multipart-sequential-kitchen-rule",
-                severity=Severity.WARNING,
-                step_name="kitchen_rules",
-                message=(
-                    "Recipe uses make-plan or rectify but kitchen_rules do not contain "
-                    "a sequential execution constraint. Without it, agents may "
-                    "batch-verify all parts before "
-                    "implementing any. Add a rule such as: 'SEQUENTIAL EXECUTION: complete full "
-                    "cycle per part before advancing.'"
-                ),
-            )
-        )
-
-    next_or_done = wf.steps.get("next_or_done")
-    if next_or_done is not None and next_or_done.on_result is not None:
-        # Legacy format: field/routes dict with explicit "more_parts" → any step
-        more_parts_target = next_or_done.on_result.routes.get("more_parts")
-        has_more_parts_route = more_parts_target is not None and more_parts_target in wf.steps
-        # Predicate format: condition with "more_parts" in the when clause routing to any step
-        if not has_more_parts_route:
-            has_more_parts_route = any(
-                cond.route in wf.steps and cond.when is not None and "more_parts" in cond.when
-                for cond in next_or_done.on_result.conditions
-            )
-        if not has_more_parts_route:
-            findings.append(
-                RuleFinding(
-                    rule="multipart-route-back",
-                    severity=Severity.ERROR,
-                    step_name="next_or_done",
-                    message=(
-                        "Recipe uses make-plan or rectify but next_or_done does not route "
-                        "'more_parts' back to a recipe step. Sequential part processing requires "
-                        "a more_parts → <loop-back-step> condition in the on_result routes."
-                    ),
-                )
-            )
-
-    return findings
-
-
-@semantic_rule(
-    name="merge-cleanup-uncaptured",
-    description="merge_worktree steps should capture cleanup_succeeded to track orphaned results.",
-    severity=Severity.WARNING,
-)
-def _check_merge_cleanup_captured(ctx: ValidationContext) -> list[RuleFinding]:
-    wf = ctx.recipe
-    findings: list[RuleFinding] = []
-
-    for step_name, step in wf.steps.items():
-        if step.tool != "merge_worktree":
-            continue
-        # Check whether any capture value references cleanup_succeeded
-        captures_cleanup = any(
-            "result.cleanup_succeeded" in str(v) for v in (step.capture or {}).values()
-        ) or any("result.cleanup_succeeded" in str(v) for v in (step.capture_list or {}).values())
-        if not captures_cleanup:
-            findings.append(
-                RuleFinding(
-                    rule="merge-cleanup-uncaptured",
-                    severity=Severity.WARNING,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' calls merge_worktree but does not capture "
-                        f"'cleanup_succeeded'. Add a capture entry such as "
-                        f'cleanup_ok: "${{{{ result.cleanup_succeeded }}}}" '
-                        f"so cleanup failures (orphaned worktree/branch) are not silently ignored."
-                    ),
-                )
-            )
-
-    return findings
-
-
-@semantic_rule(
-    name="stale-ref-after-merge",
-    description=(
-        "A context variable sourced from a worktree path or branch name is consumed "
-        "by a step that executes after merge_worktree (or remove_clone), which deletes "
-        "the resource the variable refers to."
-    ),
-    severity=Severity.WARNING,
-)
-def _check_stale_ref_after_merge(ctx: ValidationContext) -> list[RuleFinding]:
-    return [
-        RuleFinding(
-            rule="stale-ref-after-merge",
-            severity=Severity.WARNING,
-            step_name=w.step_name,
-            message=w.message,
-        )
-        for w in ctx.dataflow.warnings
-        if w.code == "REF_INVALIDATED"
-    ]
-
-
-@semantic_rule(
-    name="uncaptured-handoff-consumer",
-    description=(
-        "Skill with no declared outputs before a consumer with unwired optional file-path inputs"
-    ),
-    severity=Severity.WARNING,
-)
-def _check_uncaptured_handoff_consumer(ctx: ValidationContext) -> list[RuleFinding]:
-    """Warning when an outputs:[] skill is immediately followed by a consumer with optional
-    file-path inputs that are not wired via ${{ context.* }} references.
-
-    This is the consumer-side complement to the implicit-handoff rule (which is producer-side).
-    Together they form a bidirectional handoff enforcement pair.
-    """
-    try:
-        manifest = load_bundled_manifest()
-    except Exception:
-        logger.warning(
-            "uncaptured-handoff-consumer: failed to load skill_contracts.yaml; skipping"
-        )
-        return []
-
-    findings: list[RuleFinding] = []
-
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool not in SKILL_TOOLS:
-            continue
-
-        producer_skill = resolve_skill_name(step.with_args.get("skill_command", ""))
-        if not producer_skill:
-            continue
-
-        producer_contract = get_skill_contract(producer_skill, manifest)
-        if producer_contract is None:
-            continue
-        if producer_contract.outputs:
-            continue  # has declared outputs — implicit-handoff rule covers this
-
-        # producer declares outputs: [] — check successors for unsatisfied file-path inputs
-        for successor_name in ctx.step_graph.get(step_name, set()):
-            successor_step = ctx.recipe.steps.get(successor_name)
-            if successor_step is None or successor_step.tool not in SKILL_TOOLS:
-                continue
-
-            consumer_skill = resolve_skill_name(successor_step.with_args.get("skill_command", ""))
-            if not consumer_skill:
-                continue
-
-            consumer_contract = get_skill_contract(consumer_skill, manifest)
-            if consumer_contract is None:
-                continue
-
-            file_path_inputs = [
-                inp
-                for inp in consumer_contract.inputs
-                if inp.type in ("file_path", "directory_path") and not inp.required
-            ]
-            if not file_path_inputs:
-                continue
-
-            skill_cmd = successor_step.with_args.get("skill_command", "")
-            unwired = [inp for inp in file_path_inputs if f"context.{inp.name}" not in skill_cmd]
-            if not unwired:
-                continue  # all file-path inputs wired via context refs
-
-            input_names = ", ".join(inp.name for inp in unwired)
-            findings.append(
-                RuleFinding(
-                    rule="uncaptured-handoff-consumer",
-                    severity=Severity.WARNING,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' invokes '{producer_skill}' which declares no "
-                        f"structured outputs (outputs: []). Its successor '{successor_name}' "
-                        f"('{consumer_skill}') has optional file-path inputs ({input_names}) "
-                        f"not provided via ${{{{ context.* }}}} references. If '{producer_skill}' "
-                        f"writes files consumed by '{consumer_skill}', add output emission to "
-                        f"the skill and a capture: block on this step."
-                    ),
-                )
-            )
-
-    return findings
-
-
-def _get_provided_args(with_args: dict) -> set[str]:
-    _nested = with_args.get("args")
-    nested_args: set[str] = set(_nested.keys()) if isinstance(_nested, dict) else set()
-    top_level_args = set(with_args.keys()) - {"callable", "timeout", "args"}
-    return nested_args | top_level_args
-
-
-def _get_args_values(with_args: dict) -> dict[str, object]:
-    result: dict[str, object] = {}
-    _nested = with_args.get("args")
-    if isinstance(_nested, dict):
-        result.update(_nested)
-    for key, val in with_args.items():
-        if key not in {"callable", "timeout", "args"}:
-            result[key] = val
-    return result
-
-
-@semantic_rule(
-    name="missing-callable-input",
-    description="run_python steps must pass all required inputs declared in callable contract",
-    severity=Severity.ERROR,
-)
-def _check_missing_callable_input(ctx: ValidationContext) -> list[RuleFinding]:
-    findings = []
-    manifest = load_bundled_manifest()
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "run_python":
-            continue
-        callable_path = step.with_args.get("callable", "")
-        if not callable_path:
-            continue
-        contract = get_callable_contract(callable_path, manifest)
-        if contract is None:
-            continue
-        required_inputs = {inp.name for inp in contract.inputs if inp.required}
-        provided_args = _get_provided_args(step.with_args)
-        missing = required_inputs - provided_args
-        for arg_name in sorted(missing):
-            findings.append(
-                RuleFinding(
-                    rule="missing-callable-input",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' calls '{callable_path}' but does not pass "
-                        f"required input '{arg_name}'. Add it to the step's args block."
-                    ),
-                )
-            )
-    return findings
-
-
-@semantic_rule(
-    name="callable-signature-mismatch",
-    description="run_python step args keys must match the callable's function signature",
-    severity=Severity.ERROR,
-)
-def _check_callable_signature_mismatch(ctx: ValidationContext) -> list[RuleFinding]:
-    findings = []
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "run_python":
-            continue
-        callable_path = step.with_args.get("callable", "")
-        if not callable_path:
-            continue
-        try:
-            module_path, attr_name = callable_path.rsplit(".", 1)
-            mod = importlib.import_module(module_path)
-            func = getattr(mod, attr_name)
-            sig = inspect.signature(func)
-        except (ImportError, AttributeError, ValueError):
-            continue
-        if any(p.kind == inspect.Parameter.VAR_KEYWORD for p in sig.parameters.values()):
-            continue
-        valid_params = set(sig.parameters.keys())
-        provided_args = _get_provided_args(step.with_args)
-        invalid = provided_args - valid_params
-        for arg_name in sorted(invalid):
-            findings.append(
-                RuleFinding(
-                    rule="callable-signature-mismatch",
-                    severity=Severity.ERROR,
-                    step_name=step_name,
-                    message=(
-                        f"Step '{step_name}' passes arg '{arg_name}' to '{callable_path}' "
-                        f"but the callable does not accept that parameter."
-                    ),
-                )
-            )
-    return findings
-
-
-@semantic_rule(
-    name="downstream-context-gap",
-    description=(
-        "A step's skill_command references ${{ context.X }} but X is not produced "
-        "by any prior step capture, not declared as a recipe input, and not in the "
-        "recipe's ambient context. The variable is unreachable."
-    ),
-    severity=Severity.WARNING,
-)
-def _check_downstream_context_completeness(ctx: ValidationContext) -> list[RuleFinding]:
-    """Check that context variables referenced in skill_command are actually wired.
-
-    For each run_skill step, extract all ${{ context.X }} references from
-    skill_command and verify that X is either:
-    (a) produced by a prior step's capture: block (result.X wired to context.X), or
-    (b) declared as a top-level recipe input.
-    """
-    findings: list[RuleFinding] = []
-
-    produced_context: set[str] = set()
-    recipe_inputs: set[str] = (
-        set(ctx.recipe.ingredients.keys()) if ctx.recipe.ingredients else set()
-    )
-
-    ordered_steps = list(ctx.recipe.steps.keys())
-
-    for step_name in ordered_steps:
-        step = ctx.recipe.steps[step_name]
-
-        # Check references first so a step's own captures cannot satisfy its own refs.
-        # Captures become available to *subsequent* steps, not to the current step.
-        if step.tool in SKILL_TOOLS:
-            skill_cmd = (step.with_args or {}).get("skill_command", "")
-            if isinstance(skill_cmd, str):
-                for ref in _CONTEXT_REF_RE.findall(skill_cmd):
-                    if ref in produced_context or ref in recipe_inputs:
-                        continue
-                    findings.append(
-                        RuleFinding(
-                            rule="downstream-context-gap",
-                            severity=Severity.WARNING,
-                            step_name=step_name,
-                            message=(
-                                f"Step '{step_name}' references ${{{{ context.{ref}}}}} in its "
-                                f"skill_command but '{ref}' is not produced by any prior step's "
-                                f"capture block and is not a recipe input."
-                                " The variable is unreachable."
-                            ),
-                        )
-                    )
-
-        if step.capture or step.capture_list:
-            for ctx_var in step.capture or {}:
-                produced_context.add(ctx_var)
-            for ctx_var in step.capture_list or {}:
-                produced_context.add(ctx_var)
-
-    return findings
-
-
-@semantic_rule(
-    name="nullable-optional-context-ref",
-    description=(
-        "run_python steps must not pass optional_context_refs values to non-nullable "
-        "callable inputs without null coercion in the callable"
-    ),
-    severity=Severity.ERROR,
-)
-def _check_nullable_optional_context_ref(ctx: ValidationContext) -> list[RuleFinding]:
-    findings = []
-    manifest = load_bundled_manifest()
-    for step_name, step in ctx.recipe.steps.items():
-        if step.tool != "run_python":
-            continue
-        callable_path = step.with_args.get("callable", "")
-        if not callable_path:
-            continue
-        contract = get_callable_contract(callable_path, manifest)
-        if contract is None:
-            continue
-        non_nullable = {inp.name for inp in contract.inputs if not inp.nullable}
-        if not non_nullable:
-            continue
-        optional_refs = set(step.optional_context_refs)
-        if not optional_refs:
-            continue
-        args_values = _get_args_values(step.with_args)
-        for inp_name in sorted(non_nullable):
-            arg_val = args_values.get(inp_name)
-            if not isinstance(arg_val, str):
-                continue
-            for ref in _CONTEXT_REF_RE.findall(arg_val):
-                if ref in optional_refs:
-                    findings.append(
-                        RuleFinding(
-                            rule="nullable-optional-context-ref",
-                            severity=Severity.ERROR,
-                            step_name=step_name,
-                            message=(
-                                f"Step '{step_name}' passes optional context ref '{ref}' "
-                                f"to non-nullable input '{inp_name}' of callable "
-                                f"'{callable_path}'. Add null coercion in the callable "
-                                f"or remove from optional_context_refs."
-                            ),
-                        )
-                    )
     return findings
