+++ tests/_test_filter.py
@@ -359,13 +359,22 @@ MODULE_CASCADE_RECIPE: dict[str, frozenset[str]] = {
     "rules_bypass": frozenset({"recipe"}),
     "rules_campaign": frozenset({"recipe"}),
     "rules_ci": frozenset({"recipe"}),
+    "rules_ci_conflict": frozenset({"recipe"}),
+    "rules_ci_guards": frozenset({"recipe"}),
+    "rules_ci_merge_queue": frozenset({"recipe"}),
     "rules_clone": frozenset({"recipe"}),
     "rules_cmd": frozenset({"recipe"}),
     "rules_contracts": frozenset({"recipe"}),
     "rules_dataflow": frozenset({"recipe"}),
+    "rules_dataflow_callable": frozenset({"recipe"}),
+    "rules_dataflow_handoff": frozenset({"recipe"}),
+    "rules_dataflow_multipart": frozenset({"recipe"}),
     "rules_features": frozenset({"recipe"}),
     "rules_fixing": frozenset({"recipe"}),
     "rules_graph": frozenset({"recipe"}),
+    "rules_graph_output": frozenset({"recipe"}),
+    "rules_graph_review": frozenset({"recipe"}),
+    "rules_graph_routes": frozenset({"recipe"}),
     "rules_inline_script": frozenset({"recipe"}),
     "rules_inputs": frozenset({"recipe"}),
     "rules_isolation": frozenset({"recipe"}),
